Conditional execution of an example based on the capture of a tag
Exported the clipboard's state through the 'options' stack. From that, a
new concern named 'Conditional' can mark an example to be skipped if a
particular tag has a non-empty string value in the clipboard.

This introduces three new and mutually exclusive options: +on, +if,
+unless.

All of them take the name of a tag: +on and +if will execute the example
if the given tag has a non-empty string; +unless will execute it if it
has an empty string.