Auto hide the non-printable replacement in the enhanced diff
If a particular non-printable character *is not* present in neither the
expected output, the got output nor in the captured tags, that
particular character will not be shown in the description that explains
that some characters are begin been replaced by printable characters.

If no character is replaced, hide the full description. Closes #58.