Warn about non-printable char replaced that could be confused
If a non-printable char is replaced by X and X was before in the
original string, the character X could be confused by the literal X or
the non-printable char.

If this situation happens, warn the user. Closes #30.

In addition to this, do not replace empty lines with ^n. Instead,
replace any new line with ^n and only in the captured tags:
empty lines can be seen perfectly fine in the expected/got outputs
without any help.