Warn if RegexCache is used or file lock isn't supported
Warn if the RegexCache is used: it is still a very experimental feature
that tries to reduce the time of the regex compilation and uses very
dark and undocumented implementation details of the 're' module.

Also, warn if the current OS or file system does not support file locks,
required by the cache. In that case, disable the cache too and prevent
any race condition.

Closes #90