Logs in executor.py (draft). This introduces race conds, ignores concerns
Because the execution is made in parallel by different processes and
because the logging.Logger class handles only thread-level mutual
exclusions, the current impl has several race conditions.

This is because at the end we are not using the concerns interface and
we are not sending the messages/logs to the Progress concern.

This will change very soon.