Send the formatted logs to Progress (do not format there, small incompatible change)
Use Progress' event method to print the already formatted logs from the
new log system.

Before this commit, Progress was doing a rudimentary formatting before
printing. Now this is not available anymore which introduce a small
change in the behaviour of Progress' event but not at the API level.

This ensures that the logs from multiple jobs (processes) will be
written in the terminal/console without clashing each other (Progress
responsibility).

At the same time, the log system is not thread *unsafe*, relaying in
Progress to guaranty the thread and process *safety*.

Because the Progress module is not loaded from the begin, there is a
small time window where the log system will still be printing to the
terminal/console directly (but it is ok because there is only one
thread/job running).

When byexample is initialized (init function), the log system gets
connected with the Concern's system and the logs will be sent to
Progress' event (to all the event methods of all the Concerns to be more
precise).