PHP interpreter (experimental)
Prompts:
 php>
 ...>

The interpreter (php -a) does not turn off the echo so we use the
experimental _get_output_echo_filtered.
This forces use to use +term=ansi and a larger +geometry by default.

The interpreter does not support canceling an action nor disabling the
readline feature.

It has a builtin pretty print with var_dump and print_r but it does not
print anything by default. One of those must be called explicitly.