Extend _expect_prompt to expect (wait) for something else
_expect_prompt can optionally wait for an arbitrary regex in addition to
the prompt.

The idea is that we can wait for things *before* a prompt like a string
product from the partially executed example that may indicate that the
user should *type* something.

In this case, _expect_prompt must return earlier so we have the chance
to type something *before* the next prompt, otherwise we would always
timeout.