Extract inputs and inputs' prefixes from the expected str
Extract the inputs and the prefixes. Request a minimum of prefix only if
byexample is out of sync (see below). Also, for each input generate the
literals and whitespaces as usual (see below).

In sync concept: don't require a prefix when it is not needed.

At the begin, byexample will be in sync with the running example
and its output so there is no need to have a large prefix as
it knows that the small (even empty) prefix will be enough.

This is because a prefix gets shorter because a tag gets in the
middle or because an input gets in the middle.

In the first case byexample loose the synchronization and requires a
minimum amount of prefix.

In the second case, byexample knows that a previous input was typed
before so it was in sync then and it will be in sync now (for the
current input) so a prefix is not needed.

In any case byexample always try to build a prefix because it ensures
that it will not start typing when it shouldn't. However, and to avoid
mismatches, the prefix length is bounded.

Generate literals concept:

Do not generate input tokens only. Instead, generate from
them literals and whitespace tokens too.

This will allow the parser to process the expected string as usual,
processing the inputs as a totally separated branch.

Ensure that those literals are not used as prefixes because they are not
actually read from the example, just we fake that.