Rename last_output to output_between_prompts
The name 'output_between_prompts' is much specific than 'last_output':
we collect the output after each prompt match (what for the
interpreter's point of view is the text 'before' the match) and because
we get the output between prompt matches, 'output_between_prompts' is
better.

This highlight the fact that when _expect_prompt stops earlier due a
match of 'earlier_re', the last item in 'output_between_prompts' is no
necessary the output between prompts.