Deleted : from the tests, its escaped version changed in Python 3.8
Before Python 3.8 the colon (:) was a "future reserved" symbol and
re.escape escaped it (\\:). In Python 3.8 this changed and the escaped
version of the symbol is the symbol itself.

Several other symbols have the same behaviour (=, -, ...) so for the
sake of the simplicity, the : (and the rest) is forbidden from the tests
to have a deterministic output across different Python versions.