Allow force the echo filtering
This can be controlled by the +force-echo-filtering option available to
all the interpreters (but ignored by the ones that already have that
behavior by default).

The options can be set from the command line (with -o) to act globally
or per example.

It's an experimental feature that will:
 - break if no echo is read (the filter will eat the first line of the
output)
 - force a terminal emulation (like +term=ansi) and it will have the
same effects and limits including the terminal size (+geometry)