Find examples in Python files only in docstrings.
The zone finder for Python files (.py) is much smarter and only finds
zones and examples inside docstrings.

These docstrings follow the definition of docstring of Python: a string
that it is immediately after a function or class definition (see
https://docs.python.org/3/library/ast.html).

Because it was tradition to put some initialization code in a string at
the module level without being a module's docstring, the zone finder
will also find zones and examples there (but only in the module's
strings).

Without this change, the previous finder was too naive and it could
confuse a normal string with a docstring breaking the finder.