Force 'fork' as the multiprocessing start method
'fork' is the default in Unix and it was the default for macOS until
Python 3.8.

With this start method, objects from the parent are copied bit-by-bit
into the subprocesses' memory so they are an identical copy of its
parent.

The other methods available are 'spawn' and 'forkserver' that requires a
serialization of these objects using pickle.

Unfortunately these objects are quite complex and contain
non-serializable attributes so 'fork' is the only method available.

Note: the macOS support is still a little fragile: we use 'fork' so
byexample will work in Python 3.8 but the 'fork' in macOS has some
esoteric problems including crashes. It is a limitation of the OS and
the provided library, not of Python or byexample.

See https://bugs.python.org/issue33725