Issue #144: use home-made regex module as replacement for re
Use a home-made regex module to replace Python's re. This home-made
serves as a layer to "hide" the used of re: the only
purpose is to serve as a thin layer between client code and the real
regex engine.

Supports only compile and escape to promote the use the regex/pattern
objects instead of the global functions search/match/find/sub/...

This layer can be used later to do advanced cache or to use a different
regex engine.

This commit was originally development for #128.

(cherry picked from commit 5de8a41e026a2c6419853959271e59bfde93e83b)