Use regex module as replacement for re
Use a home-made regex module to replace Python's re. This home-made
delegates on the third-party regex module the real work: the only
purpose is to serve as a thin layer between client code and the real
regex engine.

Supports only compile and escape to promote the use the regex/pattern
objects instead of the global functions search/match/find/sub/...