Make extend_option_parser to receive a ctx (break API)
Allow the method `extend_option_parser` of `ExtendOptionParserMixin` (used
by `Concern` and `ExampleParser`) to receive a `ctx` (arbitrary, possibly
empty, dictionary/object).

This breaks the API because `ExampleParser` will pass the `ctx` of
`get_extended_option_parser` to `extend_option_parser` and if the module
(`Concern` or `ExampleParser`) does not accept that it will crash.

The `ctx` will allow to pass a state to `extend_option_parser` explicitly
so it will not required to it to store it in the instance (self) so
we'll be able to make the `get_extended_option_parser` and
`extend_option_parser` class-method (stateless)