Split init in two: harvester and executor are per worker (MAY break)
The initialization process (drove by `init()`) was split in two: the
part for initializing `byexample` as a "process" (`init_byexample`) and
the part for initializing a worker (`init_worker`).

The harvester and the executor objects are now per worker so they are
created multiple times (once per worker) and they are not shared
anymore.

The `init_worker` recreates the registry: a new registry is build
recreating the objects inside of it (parsers, delimiters, runners,
finders and concerns).

For N workers, those objects are created N+1 times: one in the context
of `init_byexample` and N in the context of `init_worker` (once per worker).

This makes the architecture more multithread-friendly but introduces
race conditions in the Progress module that relays on a mutex created in
the constructor.

This MAY breaks the API because, like Progress (a concern), a
third-party module may relay in that the constructor is called once.