Github Actions replacement for Travis CI (macos not supported)
Moved from Travis CI to Github Actions. All except the tests for macos
were ported.

For macos there is an issue where the 'echo' cannot turned off, similar
to what happen in #116. It is not clear if it is a problem of the OS,
the whay that the interpreters are installed in the testing environment,
a bug of Pexpect or a combination of.