Fake bump to 9.2.4 to override 9.2.3
The 9.2.3 was pushed to master by mistake. Master contains code not
ready to be published so a special branch was made from 9.2.2 and
a 9.2.4 was created from that.

The 9.2.3 should not break any real code but introduces a dependency
(regex module) that may not be easily installed and it is planned for
10.x.x.