Fake bump to 9.2.5 to override 9.2.3 and 9.2.4
The 9.2.3 was pushed to master by mistake. Master contains code not
ready to be published so a special branch was made from 9.2.2 and
a 9.2.4 was created from that.

Unfortunatelly the branch was also wrong and it was took one commit
ahead of 9.2.2 so a second branch was made (hopefully right this time)
and a 9.2.5 was created from that.

The 9.2.3/9.2.4 should not break any real code but introduce a dependency
(regex module) that may not be easily installed and it is planned
for 10.x.x.