Issue #1: support profile byexample in multithreading
The profiler functions (profile_ctx) will use a thread-local storage to
track the call trace separately from the call traces of other threads.

To synchronize the write to disk, the threads will use a lock and a
queue: if the lock is free, it is acquired by the thread and writes to
disk; if it is not free, the trace will be put in the queue.

Any thread that acquires the lock will also pop and print the messages
from the queue.

To ensure that no trace was left in the queue, a _flush_traces function
will be called at exit.