Issue #1: instrumented byexample
Focus in the most top functions (executor, parser, finder, init) and
then dig into the most likely and slow subfunctions.

With this we can have evidence of which parts are slow and which aren't
and the ratios between instrumented parts and non-instrumented parts can
be used for regression tests.

Tests show that:
 - pexpect's expect is really slow: when executing the examples, it
takes 46% of the whole execution time of byexample
 - the spawn of the interpreters takes 16% (12% due expect)
 - initialization of the interpreters takes 18% (16% are from the spawn)
 - shutdown of the interpreters takes 14%
 - finding related stuff only takes 4% and init only takes 0.6%