Issue #131: make log system thread safe; remove queue
The use of a queue made the log system to emit the messages out of order
with respect the ones emitted by the concerns by themselves (like
`Progress` printing stuff).

So the queue impl was removed.

Instead, to make the log system thread safe, `XStreamHandler` uses a lock
to protect only the `StreamHandler` (stderr) access only. It relays on
concerns' thread-safety for the rest.

For that to make sense, `XStreamHandler` now uses a concerns set per
thread: on init of each thread, the thread must configure its own
thread-specific system with its own concerns set.

This "thread-specific concerns set" concept is not new. Each thread
(worker) has its own copy of the harvester, executor and concerns. This
commits only ensures that that thread-specific concerns sets is
available to the `XStreamHandler`.