Issue #147: make accessible the current options from example
The example has an 'options' attributes that contains the options of the
example but they don't represent the whole "stack" of options.

This is defined only in runtime by the FileExecutor.

Now this is set and made accessible just before the example is executed
and removed at the end of its execution (before parsing the next
example) so the runner, the differ and the concerns will have access
without passing explicitly in each call the 'options'.