Issue #114: call runner.reset() if possible, fallback to shutdown()
If we don't want to force a shutdown and the runner supports a reset(),
do it instead of a shutdown.

A reseted runner will still be alive and a subsequent initialization
will be skipped.

This allows the runner to implement a reset of the interpreter's state
instead of a full shutdown/re-spawn.

Documented runner's initialize(), reset() and shutdown() and refactored
FileExecutor's initialize_runners() and shutdown_runners() (now called
reset_runners()).

This should backward compatible.

FileExecutor for now it always force a shutdown so the whole thing works
as before.

Also it is missing the shutdown of the still alive runners.