Issue #143: Config 'job_number' names the worker number or __main__
If the workers/jobs were not spawned yet, the Config object will have
'__main__' as the value of 'job_number'.

In each worker/job this value is patched to have the job's number so an
user code (like a Concern or Runner) can know when it is the main thread
or in a worker.