Issue #143: polymorphic concurrency objects
The byexample.concurrency module offer a way to load three classes
Process, Manager and Queue that are polymorphic based on the concurrency
model.

For multiprocessing model, the three classes come from the
multiprocessing module.

For multithreading model, the three classes come from the
multiprocessing.dummy module. The Manager class it was extended to
support context manager syntax.

For singlethreading model, the three classes are like in the
multithreading model but we may optimize them later to avoid unnecessary
locks like using _DummyLock.