Issue #143: use sharer and namespace for concurrency (API change)
Jobs now load the concurrency model (only multithreading for now) and
creates from there a Manager instance called 'sharer'.

A namespace object 'ns' is used to hold things that the Runners,
Parsers, Concerns would like to share among the workers.

This things must be created by the 'sharer' to ensure that they can be
shared among the workers.

It is up to the caller to also share synchronization mechanisms like
locks and use them to protect the access of the shared objects.

'ns' is read-write during the first phase (main thread) but it becomes
read-only when the workers start. This ensures that it is a thread safe
object itself.

'sharer' becomes None when the workers start because the shared objects
can be created only in the first phase.

The current implementation supports only multithreading with a possible
optimization to singlethreading when the number of jobs is 1.

The multiprocessing model is not supported due the following issues:
 - the 'ns' read-only object is a namedtuple and cannot be pickled for
sharing. In Python 3.6 we could replace 'ns' by a Manager.Namespace
which should be shareable but we may loose the protection of being
read-only.
 - the cfg.options cannot be pickled (due problems with Argparse)

Other issues may be lurking underground.

Implementers of Runners, Parsers, Finders, Concerns are required to use
'sharer' to create shared objects and store them in the 'ns' object
during the first phase and retrieve them from 'ns' after the workers
were spawned.

Other shared things like class attributes are not supported (they will
work in the multithreading model but they will break in the
multiprocessing model)