Issue #154: cfg tuples not allowed; 'ns' is an exception
A tuple object is constant but its contained object may not so a tuple
is not longer allowed by Config to be a valid element.

An exception to this is 'ns' which it is namespace/namedtuple.

Currently this exception may lead to a race condition if the user is not
careful enough.