Issue #154: make 'ns' shareable and immutable in the workers
In the main thread, 'ns' is a NS object where the runners, parsers,
concerns may add attributes.

At the end of the initialization, the 'ns' object is transformed into a
shareable Namespace object which can be shared among the workers in any
concurrency model (multithreading or multiprocessing).

So far 'ns' is mutable but before starting executing tests, the 'ns'
object is turned into an immutable namedtuple during the Config copy
in each worker/job.

This makes impossible to use 'ns' as a way to share things among the
workers. Instead the workers need to user 'ns' attributes for that.