Issue-158: remove any trailing cookie inserted by _filter_echo
_filter_echo inserts a fake prompt (aka cookie) between the lines
to reconstruct the original output obtained from pexpect to pass it
later to an ANSI terminal emulator and execute the ANSI escape sequences
that should put the prompts at the begin of each line to be filtered
later.

But the assumption that the interpreter in the original output printed
its prompt at the begin of the line (in the last line) it may be
incorrect.

Cling for example prints its prompt immediately after the printed output
if this one does not print a newline explicitly.

Ex:
[cling]$ std::cout << 1;
1[cling]$

In contrast with:
[cling]$ std::cout << 1 << std::endl;
1
[cling]$

This lack of newline makes _filter_echo ineffective. See issue #158.

This commit fixes that filtering the cookie from the end of each line.