Issue #129: log shutdown failures and call terminate not matter what
A close() may fail if the runner closed due the previous sent EOF; a
terminate() may fail if the runner closed due the previous close() call.

In both cases, log a debug message but call the two things
unconditionally.

If the runner is still alive, log a warning. It could be an error
however the problem is more likely to be a timing issue where the
runner is shutting down (too slow).

If we have _shutdown_interpreter to run in background we could change
the impl to do a more synchronous wait without introducing a noticeable
delay.