Three user abort types: soft, without join, hard.
When the user sends a Ctrl-C (SIGINT / KeyboardInterrupt), no more files
are enqueued and byexample will finish when the current active jobs
finish.

A second Ctrl-C will make byexample to not wait for the current active
jobs to finish. The current supported implementation of the jobs is
using multithreading and because the threads are not daemons, the
byexample process will block until all the threads finish.

However this is not guaranteed and may change in the future.

In any case, byexample releases the protection against KeyboardInterrupt
signals so subsequent Ctrl-C will make the current thread(s) to start
"crashing" with KeyboardInterrupt exceptions.

Added some tests about this.