Issue-172: allow cache as constant methods with args
Despite receiving arguments, the method must guarantee to return the
same value always. Only then the constant decorator makes sense.

If the method may return different values for different inputs
(arguments), it would be better to use `functools.lru_cache`.

The advantage of `constant` over `functools.lru_cache` is it is
thread-safe and `byexample`'s concurrency model is aware of it.