Issue-176: fix edge rc on recovering/interpreter closed
A possible edge race condition between an example that it is hung,
byexample sending an abort (ctrl-C) but in that moment the example
finishes and the signal is received by the interpreter that finishes
also leading to an InterpreterClosedUnexpectedly.

We cannot "fix" the RC but we can handle the case and print a message
instead of an exception.