Issue-176: optionally delay after prompt to capture more text
Some interpreter may have a race condition between the output generated
by the example and the prompt printed by the interpreter.

In those cases the output may appear *after* the prompt.

This not only breaks the example (the last bits of the expected output are
missing) but it also breaks the next example (unexpected bits are at the
begin of the expected output).

Assuming that the timming issue is quite small, we can add an optional
delay after the prompt to capture more data at the end and try to keep
the interpreter and byexample in sync.