Issue-175: emulate Pexpect's PtySpawn with a Subprocess
Pexpect already has a spawn object based on Subprocess: PopenSpawn.

However this class is not 100% compatible with the PtySpawn (the one
behind pexpect.spawn).

To avoid putting a lot of conditional code we create a PopenSpawnExt
layer on top of PopenSpawn and emulate all the missing behaviors (to
some extend).