Allow to set "next" version/tag for consistent-version test
The test verifies the current git tag as the current version and ensures
that it is what the documentation says.

The problem is that when we bump up the version but we still didn't tag,
the test will always fail.

With the new environment variable BYEXAMPLE_NEXT_VERSION, the test will
use that as the "current" git tag, rollbacking to "git describe" if the
env var is not set.

With this, a developer can test that the new not-tagged-yet version is
correctly reflected in all the documents.