Replace find_module with find_spec + exec_module
find_module is deprecated and 3.10 was removed. The almost immediate
replacement is call to find_spec to get a ModuleSpec and from there,
access to its loader attribute to get a Loader.

importlib's module_from_spec will create the Module and loader's
exec_module will load it (execute it).

For reference:
 - https://bugs.python.org/issue43540