Put loaded modules (plugins) into sys.modules
With this, the plugins are accessible from other modules without
requiring an explicit import. This was the default behaviour of the
previous pre-3.10 loader.

Without this, an import call would re-import the plugin making the
first and the second ones different from Python's perspective even if
they are both the same. This breaks pickle.