Factor out the import of modules from the initialization
load_modules() is split in two: one part imports and registers the
modules while the other does the proper initialization of the extension
objects (ExampleFinder, ExampleParser, ExampleRunner, ZoneDelimiter
and Concern).

The first part imports the (python) modules and registers them in
sys.modules so they objects are pickle-able and the access to them does
not require another import.

This is to maintain compatibility with the form that Python pre-3.10
used to load the modules.

Also, if a module cannot be loaded (typically due a syntax error), emit
an error by default. Adding -vvv will print the full traceback as usual.