Offer a prepare_subprocess_call to make a func multiprocessing safe
When multiprocessing uses another start method different of 'fork', it
requires that the target function and its arguments can be pickled.

In particular, multiprocessing needs to pickle them in the parent
process and un-pickle them in the child process.

Due how pickling works in Python, pickling a function only involves
storing the information needed to reload it: the bytecode is never
stored.

This makes pickling particular tricky in byexample: if a
we want to call a function that it is from one of the byexample
modules/plugins, the pickling will fail.

It will not fail when multiprocessing serialize it (dumps) but when it
deserialize it (loads) because in the child process, the byexample
modules/plugins will not be loaded in sys.modules and because they are
not in the sys.path (in principle), Python will not be able to find
them.

_prepare_subprocess_call() can wrap the target function and its arguments
and replacing the target by a _subprocess_trampoline function that will
call the former.

This _subprocess_trampoline will do all the bootstraping needed in the
child process, including the (re)loading of the modules/plugins, to make
the un-pickling work.

Most of these details are hidden from the user (plugin developer). He/she
is only required to call prepare_subprocess_call() and use the returned
target/arguments in replace of his/her.

The prepare_subprocess_call() is a partial bound function of
_prepare_subprocess_call(). The former can be obtained optionally from
the extension constructor (__init__ method of ExampleParser,
ExampleFinder, ExampleRunner, ZoneDelimiter and Concern).

The function is thread-safe and it will accessible both in the main
byexample process and in each worker thread.