After a stop-on-silence, disable it and use a normal timeout for shell's prompt recovery
If the stop-on-silence is not disabled, it will be used as the timeout
for waiting the prompt after the stop.

If the shell is too slow, it will raise an unexpected Timeout.

This slowness could happen due a slow shell, a too small stop-on-silence
timeout or because the program stopped is slow doing the clean up.

With this fix, the remaining of the default timeout is used to wait
for shell's prompt recovery.