Make the extension classes inherit from Extension and pass Config explicitly (API break)
ExampleFinder, ExampleParser, ExampleRunner, ZoneDelimiter and Concern
inherit from Extension. This super class is initialized (__init__) with
a single keyword-only parameter: cfg (the Config).

Any other keyword  parameter is accepted but discarded (to reduce the
amount of changes to the API).

Subclasses must call their parent's __init__ passing all the keyword
arguments (**kargs). This will ensure that the cfg passed will reach to
Extension.__init__.

However this is not enforced (yet).

This commit breaks the API because the constructors now require
a mandatory cfg.

So we used to have:

    parser = ExampleParser(0, 'utf8', Options())

And now we require

    from byexample.cfg import Config
    parser = ExampleParser(cfg=Config(
                verbosity=0,
                encoding='utf8',
                options=Options()
             ))

Note: it is expected that this change will have little impact on
practice as the extension classes are instantiated by byexample
and this commit takes care of passing the correct object as cfg.