Removed verbosity/encoding unused attrs from Extension classes (API break)
These are legacy and not used in all the cases. Now, the generic cfg
(Config) can carry any amount of attributes and if a class/subclass
needs one, it can query it.

This could break some extension.

In case if need, the attribute can be get from __init__ as

class MyParser(ExampleParser):
    def __init__(self, verbosity, encoding, **kargs):
        super().__init__(verbosity=verbosity, encoding=encoding, **kargs)

        self.verbosity = verbosity
        self.encoding = encoding

Or through the cfg property:

class MyParser(ExampleParser):
    def __init__(self, **kargs):
        super().__init__(**kargs)

        self.verbosity = self.cfg.verbosity
        self.encoding = self.cfg.encoding

Most of the use cases for overriding __init__ is to "capture" one or
more settings. This is not needed anymore with the cfg property.