Fail if PexpectMixin subclass is not subclass of ExampleRunner or it didn't call super __init__ (API break)
PexpectMixin was never designed as an independent mixin and small details
of the ExampleRunner are leaked.

Making PexpectMixin independent of ExampleRunner would require passing
more arguments to PexpectMixin.__init__ which will make the API break
more noticeable. (and the __init__ arguments will probably look more
awkward)

In practice, requiring PexpectMixin be inherited by a class that also
inherits from ExampleRunner is not unusual.

Requiring that PexpectMixin.__init__ be called after
ExampleRunner.__init__ is more likely to break but also it is easy to
fix.