Load an extension even if not target/language is set, but fail if still not set after extension instantiation (API break)
Before this commit, byexample ignored any class that didn't have a
target/language even if the class inherited from an extension class
like ExampleParser. Just a warning was issued.

Now, byexample will load the class anyways but it will require to have a
target/language defined. If no such attribute exists, it will raise an
exception.

The extension class can not define it but its __init__ must ensure
it is defined. This allows the extension to dynamically change its
target/language based on some configuration/options or even to disable
itself setting it to None.

ProgressBarReporter and SimpleReporter are examples of this.

This commits breaks the API (new classes are loaded) but it is unlikely
that it will break real code.