PoC of relaxing type/input regex required in term=ansi
The prefix_regex of an input item (+type feature) expects
to find the string *at the end of a line* and this will
not work if the interpreter outputs garbage.

This PoC is just to prove that, and *may be* could be reimplemented in a
proper fix but it is hard as the term=ansi emulation happens *after*
expecting the prompt/prefix_regex.

We could try to do something like the prompt_re does as this
theoretically has the same problems and limitations of prefix_regex but
still it looks really hacky.

Even after relaxing the prefix_regex, the example still fails with some
unknown whitespace added at the begin of the "? What..." line.

See test/BUG.md for a use case.
Issue found at 10.5.2.