Warn if a tab is present in the example's code (disable it with -warn-tab) (may break)
When there is a tab character in the example's code, emit a warning.
These tab characters may interfere with the interpreter/runner because
some of them execute an UI action (like autocomplete) on a tab key
stroke.

byexample configured its runners to not do such UI action however it is
not entirely possibly to dismiss the issue and because "if" it happens,
it is really hard to understand and debug it, it is better to warn about
it preventively.

This "may break" backward compatibility because it may annoy the user.
If needed, the user can set -warn-tab to disable the warning.

It is not expected to break real code as using a tab in the example's
code is unlikely and most probably a mistake.