Isolate lang-specific options from parser's default and cmd line.
Previous this commit each ExampleParser returned a default options and
the options parsed from the command line and byexample/init put those in
the main Options object making them active for all the rest of
byexample regardless of the examples' language.

This allowed one parser to change the default values for an option and
affect other parser/example of other language.

PythonParser was a case of this: when +py-doctest was enabled, the +tags
was disabled which makes sense for Python's Doctest but clearly breaks
any other language as +tags is expected to be enabled by default, not
disabled.

So the Python-specific option +py-doctest was indeed affecting
non-Python examples.

The isolation implemented makes byexample/init to store each
ExampleParser's defaults and options from command line in a separated
dictionary and indexed them by language. This
language_specific_defaults is available via the Options as another
entry and not as another options layer (they are not active)

FileExecutor is then now responsible of "activating" the lang-specific
defaults, putting them on top of the Options before executing
ExampleRunner's code or parsing/running an example.

The change should be transparent to ExampleParser and ExampleRunner but
it may be noticed by any ZoneDelimiter, ExampleFinder and in particular
to any Concern however it is belived that no real use-case exists for
requiring a Concern to have access to lang-specific options.