Turn echo off before executing each example (disabled by default)
Change the echo status to OFF on each example execution. This helps to
reduce the need of an active echo-filtering at the expense of some
additional overhead.

Because this could break users' examples (very unlikely), we set this to
False by default (no turn off is done) and we delay the default to True
(turn the echo off) for byexample 11.0.0