turn-echo-off is True by default and -x-turn-echo-off-on-spawn (True) is introduced (may break)
The -x-turn-echo-off flag introduced in 10.5.4, disabled by default, is
enabled (set to True) by default on 11.0.0. This may break existing
examples if they relay on the echo *not* being filtered.

To reduce the impact of the change, if force-echo-filtering is on, the
-x-turn-echo-off has no effect. Cases where force-echo-filtering is off
and still the echo is required are very unlikely (but
docs/advanced/echo-filtering.md is on example).

We add -x-turn-echo-off-on-spawn flag (True by default) to control the
echo filtering at the spawn of a runner.