feat: log the location of the log call
To track from where a log line is being emitted, if debug is enabled,
the log itself will contain the file path and line number where it was
emitted.
The code uses some heuristics to distinguish the point in the call stack
where the user code calls the log code. It is in that boundary where we
want to know the location of the caller.