fix: drain interpreter's out buffer after a window resize (geometry change)
On a window resize the interpreter may decide to redraw the screen which
will put nonsense on its buffer. On the next example, that output will
be mixed with the examples true output.

To prevent that, a drain is made, reading and discarding the output (if
any).

When draining the output of the interpreter, wait a full time period
(timeout) before leaving ensuring enough time to receive any extra
output.

Also reset the internal buffers in case something went into them before
the drain.