feat: on echo filtering, drain the interpreter's buffers
In some cases the previous example may not left things unread in the
interpreter's buffers. This could happen if due an echo a "fake/echoed"
prompt was read and interpreted by byexample by a real prompt leaving the
previous example's output got probably truncated.

This will probably end in an error but the issue gets worse as the
unread output is read by the next example executed, mixing output (plus
echos) all in one place. Not only it is wrong but also very confusing.

Draining the interpreter's buffers mitigates and isolates the problem
at the expense of making the execution a little slower.