refactor: linefeed/carriage return regexps in newline normalization
We need two kind of normalizations: to \n only and to \r\n only.

The former was named "universal newline" as it is known in Python
and it is needed to normalize the output coming from different
interpreters. Now it is renamed as LFCR_or_LF_REGEX

The latter normalization is required to be correctly interpreted by
termscraper's terminal. Named LF_or_CR_REGEX.