fix: add line feed into the echo-emulated typed input
byexample emulates and echo of the typed input so it appears in the
example's output (got) and it can be compared with the expected (which
contains the typed input).

However it incorrectly simulated a new line as a single carriage return (\n)
without a line feed (\r). This works with +term=dumb but not with +term=ansi.

Adding a line feed should fix the issue for +term=ansi because \r\n is
the correct and expected line ending in those cases.

It also should not break any example with +term=dumb because
that terminal emulation normalizes any kind of line ending into
a single carriage return (\n) anyways.