feat: append a newline to example's got as workaround for common mistakes
It is common to use <...> to ignore uninteresting stuff at the end of an
example but if the example does not end with a newline, the newline
between the interesting stuff and the <...> is expected and not found
which leads to an error.

There is a FAQ about this.

Since a few versions byexample ignores any newline at the end of an
example. So we append a newline: examples without a <...> will not fail
because byexample will ignore the newline and examples with <...> will
may (may not) eat it.

So the user will not have to worry about this.

Moreover, terminal emulation (+term=ansi) strips any empty
bottom line which a +term=dumb may not, so handling both cases as the
same due to the append of a \n reduced the differences between the
modes (and the expected of an example).